Refactor to allow for more than 2 services types.
Before, we had "mock" or "real", where real was the Service for the CANserver (PandaServiceImpl).
Now we should expand to handle more Service choices.
